build: enhance linter configuration with 12 new linters
Add critical and quality linters:
- zerologlint: Detect missing .Send() or .Msg() in zerolog
- testifylint: Best practices for testify assertions
- nilerr: Detect nil returns when error is not nil
- nilnil: Prevent returning both nil error and nil value
- contextcheck: Ensure proper context propagation
- errorlint: Error wrapping best practices (Go 1.13+)
- thelper: Test helpers must call t.Helper()
- unconvert: Remove unnecessary type conversions
- usestdlibvars: Use stdlib constants instead of magic values
- whitespace: Clean up unnecessary newlines (auto-fixable)
- reassign: Prevent reassigning package variables
- predeclared: Don't shadow Go predeclared identifiers

Fix 3 issues found by new linters:
- cmd/root.go: Use errors.As() for wrapped error checking
- internal/docs/generator.go: Use errors.Join() for multiple errors
- cmd/flags.go: Remove unnecessary type conversion

All tests passing (618 tests, 91.6% coverage)

+++ cmd/flags.go
@@ -178,7 +178,7 @@ func intDefault(v interface{}) int {
 		return int(t)
 	case uint:
 		// Check for overflow - uint can be larger than max int
-		maxInt := uint(^uint(0) >> 1)
+		maxInt := ^uint(0) >> 1
 		if t > maxInt {
 			log.Warn().
 				Uint("value", t).
